refactor: update project structure
Include the package module in src (makes sense, and makes uv happy).

Renames to avoid collisions:
* bank2ynab.converter.bank2ynab -> bank2ynab.converter.convert
* bank2ynab.pyw -> bank2ynab-gui.pyw
  * This was an issue with pytest thinking that the .pyw was the library
