+++ src/bank2ynab/config.py
